+++ CompassHUD.lua
@@ -57,10 +57,12 @@ local Options
 local HUD
 local timer = 0
 local groupThrottle = 0
+local gatherMateThrottle = 0
 local player = {x = 0, y = 0, angle = 0}
 local tomTomActive
 local questPointsTable = {}
 local groupPointsTable = {}
+local gatherMatePointTable = {}
 local HBDmaps = {}
 local directions = {
     [0] = {letter = "N" , main = true },
@@ -646,11 +648,37 @@ Addon.Defaults = {
         GroupPartyNameBackgroundColor = {r = 1, g = 1, b = 1, a = 1},
         GroupRaidNameBorder           = 'Blizzard Dialog Gold',
         GroupRaidNameBorderThickness  = 2.5,
-        GroupRaidNameBorderClass     = false,
+        GroupRaidNameBorderClass      = false,
         GroupRaidNameBorderColor      = {r = 255/255, g = 215/255, b = 0/255, a = 1},
         GroupRaidNameBackground       = 'Blizzard Tooltip',
         GroupRaidNameBackgroundClass  = false,
         GroupRaidNameBackgroundColor  = {r = 1, g = 1, b = 1, a = 1},
+        GatherMateEnabled            = false,
+        GatherMateRadius             = 150,
+        GatherMateInterval           = 6,
+        GatherMateOffset             = 20,
+        GatherMateScale              = 0.8,
+        GatherMateShowDistance       = false,
+        GatherMateShowTTA            = false,
+        GatherMateShowTitle          = false,
+        GatherMateDistanceOffset     = 0,
+        GatherMateTtaOffset          = 0,
+        GatherMateTitleOffset        = 0,
+        GatherMateDistanceCustomFont = false,
+        GatherMateDistanceFont       = "Friz Quadrata TT",
+        GatherMateDistanceFontSize   = 12,
+        GatherMateDistanceFontColor  = {r = 255/255, g = 215/255, b = 0/255, a = 1},
+        GatherMateDistanceFontFlags  = "",
+        GatherMateTtaCustomFont      = false,
+        GatherMateTtaFont            = "Friz Quadrata TT",
+        GatherMateTtaFontSize        = 12,
+        GatherMateTtaFontColor       = {r = 255/255, g = 215/255, b = 0/255, a = 1},
+        GatherMateTtaFontFlags       = "",
+        GatherMateTitleCustomFont    = false,
+        GatherMateTitleFont          = "Friz Quadrata TT",
+        GatherMateTitleFontSize      = 12,
+        GatherMateTitleFontColor     = {r = 255/255, g = 215/255, b = 0/255, a = 1},
+        GatherMateTitleFontFlags     = "",
     },
 }
 
@@ -2140,6 +2168,300 @@ Addon.Options = {
     },
 }
 
+local IntegrationsOptions = {
+    type = "group",
+    order = 10,
+    name ="Integrations",
+    get = function(info)
+        return Addon.db.profile[info[#info]]
+    end,
+    set = function(info, value)
+        Addon.db.profile[info[#info]] = value
+        Addon:UpdateHUDSettings()
+    end,
+    args = {},
+}
+
+local GatherMateOptions = {
+    type = "group",
+    order = 10,
+    name = "GatherMate",
+    args = {
+        GatherMateEnabled = {
+            type = "toggle",
+            name = "Enabled",
+            width = "full",
+            order = 10,
+        },
+        GatherMateRadius = {
+            type = "range",
+            order = 20,
+            name = "Scaning radius",
+            min = 1,
+            max = 2000,
+            softMin = 50,
+            softMax = 1000,
+            step = 1,
+            bigStep = 5,
+        },
+        GatherMateInterval = {
+            type = "range",
+            order = 30,
+            name = "Update throttle",
+            desc = "Interval between adding/removing nodes on the HUD based on their distance, relative to the 'Update Frequency' set on the 'General' tab.\nIf set to 6 (default) and 'Update Frequency' is set to 60 (default), then visibility check will be run 60/6 = 10 times per second.",
+            min = 1,
+            max = 100,
+            softMin = 1,
+            softMax = 10,
+            step = 1,
+        },
+        HeaderGatherMateNode = {
+            type = "header",
+            order = 50,
+            name = "Node"
+        },
+        GatherMateOffset = {
+            type = "range",
+            order = 60,
+            name = "Vertical adjustment",
+            min = -64,
+            max = 64,
+            step = 0.5,
+        },
+        GatherMateScale = {
+            type = "range",
+            order = 70,
+            name = "Scale",
+            min = 0,
+            max = 3,
+            step = 0.01,
+            isPercent = true,
+        },
+        HeaderGatherMateDistance = {
+            type = "header",
+            order = 100,
+            name = "Distance text"
+        },
+        GatherMateShowDistance = {
+            type = "toggle",
+            order = 110,
+            name = "Show",
+        },
+        GatherMateDistanceOffset = {
+            type = "range",
+            order = 120,
+            name = "Vertical adjustment",
+            min = -20,
+            max = 20,
+            step = 0.5,
+        },
+        GatherMateDistanceCustomFont = {
+            type = "toggle",
+            order = 130,
+            name = "Use custom font",
+        },
+        GatherMateDistanceFont = {
+            type = "select",
+            order = 140,
+            name = "Font",
+            width = 1,
+            dialogControl = "LSM30_Font",
+            values = AceGUIWidgetLSMlists['font'],
+            disabled = function() return not Options.GatherMateDistanceCustomFont end,
+        },
+        GatherMateDistanceFontSize = {
+            type = "range",
+            order = 150,
+            name = "Size",
+            width = 3/4,
+            min = 2,
+            max = 36,
+            step = 0.5,
+            disabled = function() return not Options.GatherMateDistanceCustomFont end,
+        },
+        GatherMateDistanceFontFlags = {
+            type = "select",
+            order = 160,
+            name = "Outline",
+            width = 3/4,
+            values = {
+                [""] = "None",
+                ["OUTLINE"] = "Normal",
+                ["THICKOUTLINE"] = "Thick",
+            },
+            disabled = function() return not Options.GatherMateDistanceCustomFont end,
+        },
+        GatherMateDistanceFontColor = {
+            type = "color",
+            order = 170,
+            name = "Color",
+            width = 1/2,
+            hasAlpha = true,
+            get = function(info)
+                local color = Addon.db.profile[info[#info]]
+                return color.r, color.g, color.b, color.a
+            end,
+            set = function (info, r, g, b, a)
+                local color = Addon.db.profile[info[#info]]
+                color.r = r
+                color.g = g
+                color.b = b
+                color.a = a
+                Addon:UpdateHUDSettings()
+            end,
+            disabled = function() return not Options.GatherMateDistanceCustomFont end,
+        },
+        HeaderGatherMateTTA = {
+            type = "header",
+            order = 200,
+            name = "Time to arrive"
+        },
+        GatherMateShowTTA = {
+            type = "toggle",
+            order = 210,
+            name = "Show",
+        },
+        GatherMateTtaOffset = {
+            type = "range",
+            order = 220,
+            name = "Vertical adjustment",
+            min = -20,
+            max = 20,
+            step = 0.5,
+        },
+        GatherMateTtaCustomFont = {
+            type = "toggle",
+            order = 230,
+            name = "Use custom font",
+        },
+        GatherMateTtaFont = {
+            type = "select",
+            order = 240,
+            name = "Font",
+            width = 1,
+            dialogControl = "LSM30_Font",
+            values = AceGUIWidgetLSMlists['font'],
+            disabled = function() return not Options.GatherMateTtaCustomFont end,
+        },
+        GatherMateTtaFontSize = {
+            type = "range",
+            order = 250,
+            name = "Size",
+            width = 3/4,
+            min = 2,
+            max = 36,
+            step = 0.5,
+            disabled = function() return not Options.GatherMateTtaCustomFont end,
+        },
+        GatherMateTtaFontFlags = {
+            type = "select",
+            order = 260,
+            name = "Outline",
+            width = 3/4,
+            values = {
+                [""] = "None",
+                ["OUTLINE"] = "Normal",
+                ["THICKOUTLINE"] = "Thick",
+            },
+            disabled = function() return not Options.GatherMateTtaCustomFont end,
+        },
+        GatherMateTtaFontColor = {
+            type = "color",
+            order = 270,
+            name = "Color",
+            width = 1/2,
+            hasAlpha = true,
+            get = function(info)
+                local color = Addon.db.profile[info[#info]]
+                return color.r, color.g, color.b, color.a
+            end,
+            set = function (info, r, g, b, a)
+                local color = Addon.db.profile[info[#info]]
+                color.r = r
+                color.g = g
+                color.b = b
+                color.a = a
+                Addon:UpdateHUDSettings()
+            end,
+            disabled = function() return not Options.GatherMateTtaCustomFont end,
+        },
+        HeaderGatherMateTitle = {
+            type = "header",
+            order = 300,
+            name = "Node name"
+        },
+        GatherMateShowTitle = {
+            type = "toggle",
+            order = 310,
+            name = "Show",
+        },
+        GatherMateTitleOffset = {
+            type = "range",
+            order = 320,
+            name = "Vertical adjustment",
+            min = -20,
+            max = 20,
+            step = 0.5,
+        },
+        GatherMateTitleCustomFont = {
+            type = "toggle",
+            order = 330,
+            name = "Use custom font",
+        },
+        GatherMateTitleFont = {
+            type = "select",
+            order = 340,
+            name = "Font",
+            width = 1,
+            dialogControl = "LSM30_Font",
+            values = AceGUIWidgetLSMlists['font'],
+            disabled = function() return not Options.GatherMateTitleCustomFont end,
+        },
+        GatherMateTitleFontSize = {
+            type = "range",
+            order = 350,
+            name = "Size",
+            width = 3/4,
+            min = 2,
+            max = 36,
+            step = 0.5,
+            disabled = function() return not Options.GatherMateTitleCustomFont end,
+        },
+        GatherMateTitleFontFlags = {
+            type = "select",
+            order = 360,
+            name = "Outline",
+            width = 3/4,
+            values = {
+                [""] = "None",
+                ["OUTLINE"] = "Normal",
+                ["THICKOUTLINE"] = "Thick",
+            },
+            disabled = function() return not Options.GatherMateTitleCustomFont end,
+        },
+        GatherMateTitleFontColor = {
+            type = "color",
+            order = 370,
+            name = "Color",
+            width = 1/2,
+            hasAlpha = true,
+            get = function(info)
+                local color = Addon.db.profile[info[#info]]
+                return color.r, color.g, color.b, color.a
+            end,
+            set = function (info, r, g, b, a)
+                local color = Addon.db.profile[info[#info]]
+                color.r = r
+                color.g = g
+                color.b = b
+                color.a = a
+                Addon:UpdateHUDSettings()
+            end,
+            disabled = function() return not Options.GatherMateTitleCustomFont end,
+        },
+    },
+}
+
 local function GetQuestPOIInfo(questID)
     local completed = IsQuestComplete(questID)
 
@@ -2417,7 +2739,7 @@ end
 local function questPointerSetTexts(frame, dt)
     if player.inInstance then return end
 
-    frame.distance = HBD:GetWorldDistance(questPointsTable[frame.questID].instance, player.x, player.y, questPointsTable[frame.questID].x, questPointsTable[frame.questID].y)
+    frame.distance = HBD:GetWorldDistance(frame.instance, player.x, player.y, frame.x, frame.y)
     if not frame.distance then
         frame:Hide()
         return
@@ -2434,7 +2756,7 @@ local function questPointerSetTexts(frame, dt)
         if not speed or speed == 0 then -- delta
             frame.oldDistance = frame.distance
             C_Timer.After(1, function()
-                local currentDistance = HBD:GetWorldDistance(questPointsTable[frame.questID].instance, player.x, player.y, questPointsTable[frame.questID].x, questPointsTable[frame.questID].y)
+                local currentDistance = HBD:GetWorldDistance(frame.instance, player.x, player.y, frame.x, frame.y)
                 if currentDistance then
                     speed = math.abs(currentDistance - frame.oldDistance)
                 else
@@ -2878,6 +3200,166 @@ local function setGroupIcons()
     setGroupStrataLevels()
 end
 
+local function updateGatherMateNode(node)
+    local scale = Options.Scale * Options.VerticalScale * Options.GatherMateScale
+    node.frame:SetSize(textureHeight * Options.GatherMateScale, textureHeight * Options.GatherMateScale)
+    local gameFontNormal = { fontColor = {}}
+    gameFontNormal.font, gameFontNormal.fontSize, gameFontNormal.fontFlags = GameFontNormal:GetFont()
+    gameFontNormal.fontColor.r, gameFontNormal.fontColor.g, gameFontNormal.fontColor.b, gameFontNormal.fontColor.a = GameFontNormal:GetTextColor()
+
+    if Options.GatherMateDistanceCustomFont then
+        local font = LSM:Fetch("font", Options.GatherMateDistanceFont)
+        node.frame.DistanceText:SetFont(font, Options.GatherMateDistanceFontSize * scale, Options.GatherMateDistanceFontFlags)
+        node.frame.DistanceText:SetTextColor(Options.GatherMateDistanceFontColor.r, Options.GatherMateDistanceFontColor.g, Options.GatherMateDistanceFontColor.b, Options.GatherMateDistanceFontColor.a)
+    else
+        node.frame.DistanceText:SetFont(gameFontNormal.font, gameFontNormal.fontSize * scale, gameFontNormal.fontFlags)
+        node.frame.DistanceText:SetTextColor(gameFontNormal.fontColor.r, gameFontNormal.fontColor.g, gameFontNormal.fontColor.b, gameFontNormal.fontColor.a)
+    end
+    if Options.GatherMateTtaCustomFont then
+        local font = LSM:Fetch("font", Options.GatherMateTtaFont)
+        node.frame.TimeText:SetFont(font, Options.GatherMateTtaFontSize * scale, Options.GatherMateTtaFontFlags)
+        node.frame.TimeText:SetTextColor(Options.GatherMateTtaFontColor.r, Options.GatherMateTtaFontColor.g, Options.GatherMateTtaFontColor.b, Options.GatherMateTtaFontColor.a)
+    else
+        node.frame.TimeText:SetFont(gameFontNormal.font, gameFontNormal.fontSize * scale, gameFontNormal.fontFlags)
+        node.frame.TimeText:SetTextColor(gameFontNormal.fontColor.r, gameFontNormal.fontColor.g, gameFontNormal.fontColor.b, gameFontNormal.fontColor.a)
+    end
+    if Options.GatherMateTitleCustomFont then
+        local font = LSM:Fetch("font", Options.GatherMateTitleFont)
+        node.frame.Title:SetFont(font, Options.GatherMateTitleFontSize * scale, Options.GatherMateTitleFontFlags)
+        node.frame.Title:SetTextColor(Options.GatherMateTitleFontColor.r, Options.GatherMateTitleFontColor.g, Options.GatherMateTitleFontColor.b, Options.GatherMateTitleFontColor.a)
+    else
+        node.frame.Title:SetFont(gameFontNormal.font, gameFontNormal.fontSize * scale, gameFontNormal.fontFlags)
+        node.frame.Title:SetTextColor(gameFontNormal.fontColor.r, gameFontNormal.fontColor.g, gameFontNormal.fontColor.b, gameFontNormal.fontColor.a)
+    end
+
+    local point = "TOP"
+    local relativePoint = "BOTTOM"
+    local distanceTextPosition = Options.GatherMateDistanceOffset
+    local timeTextPosition = - ((Options.GatherMateShowDistance and (Options.GatherMateDistanceFontSize * 1.2 * scale)) or 0) - Options.GatherMateTtaOffset
+    local titlePosition = - ((Options.GatherMateShowDistance and (Options.GatherMateDistanceFontSize * 1.2 * scale)) or 0) - ((Options.GatherMateShowTTA and (Options.GatherMateTtaFontSize * 1.2 * scale)) or 0) - Options.GatherMateTitleOffset
+    node.flipped = false
+    if Options.GatherMateOffset > 0 then
+        node.flipped = true
+        point = "BOTTOM"
+        relativePoint = "TOP"
+        distanceTextPosition = ((Options.GatherMateShowTTA and (Options.GatherMateTtaFontSize * 1.2 * scale)) or 0) + Options.GatherMateDistanceOffset - 4
+        timeTextPosition = Options.GatherMateTtaOffset - 4
+        if Options.GatherMateShowTitle then
+            titlePosition = Options.GatherMateTitleOffset - 4
+            timeTextPosition = timeTextPosition + (Options.GatherMateTitleFontSize * 1.2 * scale) - 4
+            distanceTextPosition = distanceTextPosition + (Options.GatherMateTitleFontSize * 1.2 * scale) - 4
+        end
+    end
+
+    node.frame.DistanceText:ClearAllPoints()
+    node.frame.DistanceText:SetPoint(point, node.frame, relativePoint, 0, distanceTextPosition)
+    node.frame.TimeText:ClearAllPoints()
+    node.frame.TimeText:SetPoint(point, node.frame, relativePoint, 0, timeTextPosition)
+    node.frame.Title:ClearAllPoints()
+    node.frame.Title:SetPoint(point, node.frame, relativePoint, 0, titlePosition)
+
+    node.frame.DistanceText:SetShown(Options.GatherMateShowDistance)
+    node.frame.TimeText:SetShown(Options.GatherMateShowTTA)
+    node.frame.Title:SetShown(Options.GatherMateShowTitle)
+end
+
+local function updateGatherMate()
+    for _, nodes in pairs(gatherMatePointTable) do
+        for _, node in pairs(nodes) do
+            if not Options.GatherMateEnabled then
+                node.visible = false
+                if node.frame then
+                    node.frame:Hide()
+                end
+            end
+            if node.frame then
+                updateGatherMateNode(node)
+            end
+        end
+    end
+end
+
+local function createGatherMateNode(nodeID, nodeType, table)
+    local gatherMateNode =  table.frame or CreateFrame("FRAME", nil, HUD)
+	gatherMateNode:SetSize(textureHeight, textureHeight)
+	gatherMateNode:SetPoint("CENTER");
+	gatherMateNode.texture = gatherMateNode:CreateTexture(nil, "ARTWORK")
+	gatherMateNode.texture:SetAllPoints(gatherMateNode)
+	gatherMateNode.texture:SetTexture(GatherMate2.nodeTextures[nodeType][nodeID])
+	gatherMateNode:Hide()
+    gatherMateNode.DistanceText = gatherMateNode:CreateFontString(nil, "BACKGROUND", "GameFontNormal")
+    gatherMateNode.DistanceText:SetJustifyV("TOP")
+    gatherMateNode.DistanceText:SetSize(0, 16)
+    gatherMateNode.DistanceText:SetParent(gatherMateNode)
+    gatherMateNode.TimeText = gatherMateNode:CreateFontString(nil, "BACKGROUND", "GameFontNormal")
+    gatherMateNode.TimeText:SetJustifyV("TOP")
+    gatherMateNode.TimeText:SetSize(0, 16)
+    gatherMateNode.TimeText:SetParent(gatherMateNode)
+    gatherMateNode.Title = gatherMateNode:CreateFontString(nil, "BACKGROUND", "GameFontNormal")
+    gatherMateNode.Title:SetJustifyV("TOP")
+    gatherMateNode.Title:SetSize(0, 16)
+    gatherMateNode.Title:SetParent(gatherMateNode)
+    gatherMateNode.Title:SetText(GatherMate2:GetNameForNode(nodeType, nodeID))
+    gatherMateNode.type = "GatherMate"
+    gatherMateNode.instance = table.instance
+    gatherMateNode.x = table.x
+    gatherMateNode.y = table.y
+    gatherMateNode.elapsed = 0
+    gatherMateNode:SetScript("OnUpdate", questPointerSetTexts)
+	return gatherMateNode
+end
+
+local function setGatherMateNodes()
+    if not Options.GatherMateEnabled or not GatherMate2 then return end
+    if gatherMateThrottle and gatherMateThrottle >= Options.GatherMateInterval then
+        local x, y = HBD:GetZoneCoordinatesFromWorld(player.x, player.y, player.uiMapID, false)
+        gatherMateThrottle = 0
+        for map, nodes in pairs(gatherMatePointTable) do
+            for coord, node in pairs(nodes) do
+                gatherMatePointTable[map][coord].visible = false
+            end
+        end
+		for _, db_type in pairs(GatherMate2.db_types) do
+			if GatherMate2.Visible[db_type] then
+				for coord, nodeID in GatherMate2:FindNearbyNode(player.uiMapID, x, y, db_type, Options.GatherMateRadius) do
+					if not gatherMatePointTable[player.uiMapID] then
+                        gatherMatePointTable[player.uiMapID] = {}
+                    end
+                    if not gatherMatePointTable[player.uiMapID][coord] then
+                        local xZone, yZone = GatherMate2:DecodeLoc(coord)
+                        local xWorld, yWorld = HBD:GetWorldCoordinatesFromZone(xZone, yZone, player.uiMapID)
+                        gatherMatePointTable[player.uiMapID][coord] = { visible = true, instance = player.uiMapID, x = xWorld, y = yWorld }
+                        gatherMatePointTable[player.uiMapID][coord].frame = createGatherMateNode(nodeID, db_type, gatherMatePointTable[player.uiMapID][coord])
+                        updateGatherMateNode(gatherMatePointTable[player.uiMapID][coord])
+                    end
+                    gatherMatePointTable[player.uiMapID][coord].visible = true
+   				end
+			end
+		end
+    end
+    for _, nodes in pairs(gatherMatePointTable) do
+        for _, node in pairs(nodes) do
+            local shown = false
+            if node.visible and player.angle then
+                if node.x and node.y and node.instance then
+                    local angle = player.angle - HBD:GetWorldVector(node.instance, player.x, player.y, node.x, node.y)
+                    if angle < 0 then angle = angle + (2 * PI) end
+                    if angle > PI then angle = angle - (2 * PI) end
+                    if angle then
+                        local visible = math.rad(Options.Degrees)/2
+                        if angle < visible and angle > -visible then
+                            node.frame:SetPoint("CENTER", HUD, "CENTER", texturePosition() * angle, Options.GatherMateOffset)
+                            shown = true
+                        end
+                    end
+                end
+            end
+            node.frame:SetShown(shown)
+        end
+    end
+end
+
+
 local function updateHeading()
     if HUD.heading and HUD.heading.text and player.angle then
         local heading = (360 - (player.angle * (180 / math.pi))) % 360
@@ -2913,6 +3395,7 @@ local function updateHUD(force)
     updateHeading()
     setQuestsIcons()
     setGroupIcons()
+    setGatherMateNodes()
 end
 
 local function updatePointerTextures()
@@ -2933,6 +3416,7 @@ local function onUpdate(_, elapsed)
     if timer < (1 / Options.Interval) then return end
     timer = 0
     groupThrottle = groupThrottle + 1
+    gatherMateThrottle = gatherMateThrottle + 1
     updateHUD(false)
 end
 
@@ -3009,6 +3493,10 @@ local function updateQuest(questID, x, y, uiMapID, questType, title, completed, 
     questPointsTable[questID].frame.scaleMultiplier = 1.5
     questPointsTable[questID].frame.minDistance = nil
     questPointsTable[questID].frame.QuestText:SetText(title)
+    questPointsTable[questID].frame.type = "Quest"
+    questPointsTable[questID].frame.instance = questPointsTable[questID].instance
+    questPointsTable[questID].frame.x = questPointsTable[questID].x
+    questPointsTable[questID].frame.y = questPointsTable[questID].y
     local options = Options.Pointers[questPointsTable[questID].frame.pointerType]
     if questPointsTable[questID].texture and options.worldmapTexture then
         questPointsTable[questID].frame.texture:SetTexture(questPointsTable[questID].texture)
@@ -3293,6 +3781,7 @@ function Addon:UpdateHUDSettings()
     updateHUD(true)
     updatePointerTextures()
     updateGroupTexts()
+    updateGatherMate()
     UnregisterAttributeDriver(HUD, 'state-hudvisibility')
     RegisterAttributeDriver(HUD, "state-hudvisibility", Options.Visibility)
 end
@@ -3977,9 +4466,13 @@ function Addon:ConstructDefaultsAndOptions()
     self.Options.args.Tabs.args.Pointers.args = pointersOptionsArgs
     self.Options.args.Profiles = AceDBOptions:GetOptionsTable(self.db)
     self.Options.args.Profiles.order = 80
+    IntegrationsOptions.args.GatherMate = GatherMateOptions
+    self.Options.args.AddonIntegrations = IntegrationsOptions
+
     AceConfig:RegisterOptionsTable(Const.METADATA.NAME, self.Options)
     _, Addon.categoryID = AceConfigDialog:AddToBlizOptions(Const.METADATA.NAME, nil, nil, "Tabs")
     AceConfigDialog:AddToBlizOptions(Const.METADATA.NAME, "Profiles", Const.METADATA.NAME, "Profiles")
+    AceConfigDialog:AddToBlizOptions(Const.METADATA.NAME, "Addon integrations", Const.METADATA.NAME, "AddonIntegrations")
 end
 
 function Addon:RefreshConfig()